+++ src/Components/Button.js
@@ -24,8 +24,6 @@ const Button = props => {
         'font-bold',
         'outline-none',
         'focus:outline-none',
-        'mr-1',
-        'mb-1',
         'ease-linear',
         'transition-all',
         'duration-150',