+++ src/Components/ConfirmDeleteButton.js
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
-import classNames from 'classnames';
 import PropTypes from 'prop-types';
+
+import Button from './Button';
 
 const ConfirmDeleteButton = props => {
     const {
@@ -8,9 +9,8 @@ const ConfirmDeleteButton = props => {
         confirmText,
         onClick,
         timeout,
-        className,
-        initialClasses,
-        confirmClasses,
+        initialColor,
+        confirmColor,
         ...attributes
     } = props;
 
@@ -31,32 +31,22 @@ const ConfirmDeleteButton = props => {
         }, timeout)
     }
 
-    const queryClass = classNames(
-        className,
-        initialClasses
-    )
-
-    const confirmClass = classNames(
-        className,
-        confirmClasses
-    )
-
     return (
         (waiting)
             ?
-            <button type="button" onClick={onConfirm} className={confirmClass} {...attributes}>
+            <Button color={confirmColor} onClick={onConfirm} {...attributes}>
                 <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6 inline-block -mt-1" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                     <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
                 </svg>
                 {confirmText}
-            </button>
+            </Button>
             :
-            <button type="button" onClick={onQuery} className={queryClass} {...attributes}>
+            <Button color={initialColor} onClick={onQuery} {...attributes}>
                 <svg xmlns="http://www.w3.org/2000/svg" className="h-6 w-6 inline-block -mt-1" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                     <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
                 </svg>
                 {text}
-            </button>
+            </Button>
     )
 }
 
@@ -66,17 +56,17 @@ ConfirmDeleteButton.propTypes = {
     onClick: PropTypes.func,
     timeout: PropTypes.number,
     className: PropTypes.string,
-    initialClasses: PropTypes.string,
-    confirmClasses: PropTypes.string,
+    initialColor: PropTypes.string,
+    confirmColor: PropTypes.string,
 }
 
 ConfirmDeleteButton.defaultProps = {
     text: "Delete",
     confirmText: "Delete - Are you sure?",
     timeout: 2000,
-    className: 'block mx-auto my-auto px-4 py-2 outline-none',
-    initialClasses: 'bg-red-500',
-    confirmClasses: 'bg-orange-500',
+    className: '',
+    initialColor: 'red',
+    confirmColor: 'orange',
 }
 
 export default ConfirmDeleteButton;